Change parseInt to parseFloat in the percentage eventListener.

+++ scripts.js
@@ -76,7 +76,7 @@ minus.addEventListener("click", () => {
 });
 
 percentage.addEventListener("click", () => {
-    display.textContent = parseInt(display.textContent) * 0.01;
+    display.textContent = parseFloat(display.textContent) * 0.01;
 });
 
 comma.addEventListener("click", () => {
